+++ CHANGELOG.md
@@ -110,6 +110,7 @@ _**Note**_: Breaking changes in the API
 * Fix #2678: Adds a convenience method for referring to Cache keys by namespace and name rather than item
 * Fix #2665: CustomResourceDefinitionContext.fromCrd support for v1 CustomResourceDefinition
 * Fix #2642: Update kubernetes-examples to use apps/v1 Deployment rather than extensions/v1beta1
+* Fix #2666: Align SharedIndexInformer API with Custom Resource Improvements
 
 #### New Features
 * Fix #2611: Support for Custom Resource and Custom Resource Definitions has been improved